catch up with entity_rework branch... PERFORMANCE is good

+++ Core/Source/Level_Loader.cpp
@@ -17,6 +17,7 @@
 #include <unordered_map>
 #include <unordered_set>
 
+
 std::vector<bool> LevelLoader::neighborCheck(const std::vector<std::vector<std::string>>& pixelMatrix, const std::string &pixel, int x, int y, int width, int height) {
     std::vector<std::string> friendlyPixels(1, "");
     std::vector<bool> neighbors(4, false); // {top, bottom, left, right}
@@ -52,7 +53,7 @@ std::vector<std::string> LevelLoader::neighborTag(const std::vector<std::vector<
 }
 
 int LevelLoader::getObstacleTextureIndex(const std::vector<bool>& neighbors) {
-    int numObstacles = (int)std::count(neighbors.begin(), neighbors.end(), true);
+    int numObstacles = std::count(neighbors.begin(), neighbors.end(), true);
     if (numObstacles == 1) {
         if (neighbors[0]) return 12;    // Top
         if (neighbors[1]) return 1;     // Right
@@ -158,7 +159,7 @@ std::unordered_map<std::string, int> LevelLoader::createDualGrid(
             return friendlyNeighbors[tile].count(str) ? tile : str;
         });
 
-        int numTiles = (int)std::count(tileQ.begin(), tileQ.end(), tile);
+        int numTiles = std::count(tileQ.begin(), tileQ.end(), tile);
         std::unordered_set<std::string> uniqueStrings(tileQ.begin(), tileQ.end());
 
         if (numTiles > 0) {
